[services] Separate dispatch init from executor
